+++ src/cli_builder.zig
@@ -140,7 +140,7 @@ pub fn parse(self: *Self, argv: [][:0]u8, builder_commands: ?[]const *CommandBui
     }
 
     if (parsed.options.get("version")) |_| {
-        printColored(.white, "v{s}\n", .{self.version});
+        printColored(&.{.white}, "v{s}", .{self.version});
         return;
     }
 
@@ -158,8 +158,8 @@ pub fn parse(self: *Self, argv: [][:0]u8, builder_commands: ?[]const *CommandBui
     if (command_) |cmd| {
         found_command = cmd;
     } else {
-        printColored(.red, "Unknown command: {s}\n", .{command_name});
-        printColored(.yellow, "\nAvailable commands: ", .{});
+        printColored(&.{.red}, "Unknown command: {s}\n", .{command_name});
+        printColored(&.{.yellow}, "\nAvailable commands: ", .{});
         self.listCommands();
         return;
     }
@@ -225,7 +225,7 @@ fn mergeOptions(self: *Self, command_options: ?[]const CLIOption) ![]CLIOption {
 
 pub fn showHelp(self: *Self) void {
     print("{s} \n\n", .{self.description});
-    printColored(.bold, "Usage:", .{});
+    printColored(&.{.bold}, "Usage:", .{});
     print(" {s} <command> [options]\n\n", .{self.name});
 
     self.listCommands();
@@ -236,7 +236,7 @@ pub fn showHelp(self: *Self) void {
 
 fn showCommandHelp(self: *Self, command_name: []const u8) void {
     if (self.commands.get(command_name)) |cmd| {
-        printColored(.bold, "Usage: ", .{});
+        printColored(&.{.bold}, "Usage: ", .{});
 
         if (cmd.options != null and cmd.options.?.len > 0) {
             print("{s} {s} [options]\n", .{ self.name, cmd.name });
@@ -246,26 +246,26 @@ fn showCommandHelp(self: *Self, command_name: []const u8) void {
 
         if (cmd.options) |options| {
             if (options.len > 0) {
-                printColored(.bold, "\nOptions:\n", .{});
+                printColored(&.{.bold}, "\nOptions:\n", .{});
             }
 
             for (options) |_option| {
                 const flags = if (_option.alias) |_| std.fmt.allocPrint(self.allocator, "-{s}, --{s}", .{ _option.alias.?, _option.name }) catch unreachable else std.fmt.allocPrint(self.allocator, "--{s}", .{_option.name}) catch unreachable;
                 defer self.allocator.free(flags);
 
-                print("  {s:<20} {s}\n", .{ flags, _option.description });
+                self.printOption(_option, flags);
             }
         }
     } else {
-        printColored(.red, "Unknown command: {s}\n", .{command_name});
-        printColored(.yellow, "\nAvailable commands: ", .{});
+        printColored(&.{.red}, "Unknown command: {s}\n", .{command_name});
+        printColored(&.{.yellow}, "\nAvailable commands: ", .{});
         self.listCommands();
     }
 }
 
 fn listCommands(self: *Self) void {
     if (self.commands.capacity() > 0) {
-        printColored(.bold, "Commands:\n", .{});
+        printColored(&.{.bold}, "Commands:\n", .{});
 
         var command_iterator = self.commands.iterator();
 
@@ -278,33 +278,40 @@ fn listCommands(self: *Self) void {
 
 fn showGlobalOptions(self: *Self) void {
     if (self.globalOptions.items.len > 0) {
-        printColored(.bold, "\nGlobal Options:\n", .{});
+        printColored(&.{.bold}, "\nGlobal Options:\n", .{});
 
         for (self.globalOptions.items) |_option| {
             const flags = if (_option.alias) |_| std.fmt.allocPrint(self.allocator, "-{s}, --{s}", .{ _option.alias.?, _option.name }) catch unreachable else std.fmt.allocPrint(self.allocator, "--{s}", .{_option.name}) catch unreachable;
             defer self.allocator.free(flags);
 
-            const required = if (_option.required) " (required)" else "";
-
-            var value: types.Value = .undefined;
-
-            if (_option.default) |_| switch (_option.default.?) {
-                .string => |str| value = .{ .string = str },
-                .bool => |b| value = .{ .bool = b },
-                .number => |num| value = .{ .number = num },
-                else => {},
-            };
-
-            const default_value = switch (value) {
-                .string => |default| terminal.coloredWithArgs(self.allocator, "(default: {s})", .{default}, .gray),
-                .number => |default| terminal.coloredWithArgs(self.allocator, "(default: {d})", .{default}, .gray),
-                .bool => |default| terminal.coloredWithArgs(self.allocator, "(default: {})", .{default}, .gray),
-                else => terminal.coloredWithArgs(self.allocator, "", .{}, .gray),
-            };
-
-            defer self.allocator.free(default_value);
-
-            print("  {s:<26} {s} {s}{s}{s} {s}\n", .{ flags, _option.description, color.ansiCode(.red), required, color.ansiCode(.reset), default_value });
+            self.printOption(_option, flags);
         }
     }
+}
+
+fn printOption(self: *Self, opt: CLIOption, flags: []u8) void {
+    const required = if (opt.required) " (required)" else "";
+
+    var value: types.Value = .undefined;
+
+    if (opt.default) |_| switch (opt.default.?) {
+        .string => |str| value = .{ .string = str },
+        .bool => |b| value = .{ .bool = b },
+        .number => |num| value = .{ .number = num },
+        else => {},
+    };
+
+    var default_value = switch (value) {
+        .string => |default| color.coloredWithArgs(self.allocator, "(default: {s})", &.{.gray}, .{default}) catch unreachable,
+        .number => |default| color.coloredWithArgs(self.allocator, "(default: {d})", &.{.gray}, .{default}) catch unreachable,
+        .bool => |default| color.coloredWithArgs(self.allocator, "(default: {})", &.{.gray}, .{default}) catch unreachable,
+        else => color.colored(self.allocator, "", &.{.gray}) catch unreachable,
+    };
+
+    defer default_value.deinit();
+
+    var colored_required = color.colored(self.allocator, required, &.{.red}) catch unreachable;
+    defer colored_required.deinit();
+
+    print("  {s:<26} {s} {s} {s}\n", .{ flags, opt.description, colored_required.result, default_value.result });
 }