+++ src/parser.zig
@@ -184,12 +184,12 @@ fn parseOptionValue(self: *Self, value: types.Value, option: CLIOption) types.Va
                 } else if (std.mem.eql(u8, string_value, "false")) {
                     return types.Value{ .bool = false };
                 } else {
-                    terminal.printColored(.red, "Option '{s}' expects bool but got '{s}' which can not be converted to boolean.\n", .{ option.name, @tagName(value) });
+                    terminal.printColored(&.{.red}, "Option '{s}' expects bool but got '{s}' which can not be converted to boolean.\n", .{ option.name, @tagName(value) });
                     std.debug.print("\nOption '{s}' expects bool but got: {s}.\n", .{ option.name, @tagName(value) });
                     std.process.exit(1);
                 }
             } else if (value != .bool) {
-                terminal.printColored(.red, "Option '{s}' expects bool but got: {s}.\n", .{ option.name, @tagName(value) });
+                terminal.printColored(&.{.red}, "Option '{s}' expects bool but got: {s}.\n", .{ option.name, @tagName(value) });
                 std.debug.print("\nOption '{s}' expects bool but got: {s}.\n", .{ option.name, @tagName(value) });
                 std.process.exit(1);
             }
@@ -198,13 +198,13 @@ fn parseOptionValue(self: *Self, value: types.Value, option: CLIOption) types.Va
         },
         .number => {
             if (value == .bool) {
-                terminal.printColored(.red, "Option '{s}' expects number but got: {s}.\n", .{ option.name, @tagName(value) });
+                terminal.printColored(&.{.red}, "Option '{s}' expects number but got: {s}.\n", .{ option.name, @tagName(value) });
                 std.process.exit(1);
             } else {
                 const res = parseNumber(value.string);
 
                 if (res != .number) {
-                    terminal.printColored(.red, "Option '{s}' expects number but got: {s}.\n", .{ option.name, @tagName(value) });
+                    terminal.printColored(&.{.red}, "Option '{s}' expects number but got: {s}.\n", .{ option.name, @tagName(value) });
                     std.process.exit(1);
                 }
 
@@ -215,15 +215,15 @@ fn parseOptionValue(self: *Self, value: types.Value, option: CLIOption) types.Va
             const res = parseNumber(value.string);
 
             if (res != .string) {
-                terminal.printColored(.red, "Option '{s}' expects string but got: {s}.\n", .{ option.name, @tagName(res) });
+                terminal.printColored(&.{.red}, "Option '{s}' expects string but got: {s}.\n", .{ option.name, @tagName(res) });
                 std.process.exit(1);
             }
 
             if (option.choices != null and option.choices.?.len >= 1) {
                 var choices_array: std.ArrayList([]const u8) = .empty;
                 for (option.choices.?, 0..) |choice, i| {
-                    choices_array.append(allocator, choice) catch terminal.printError("Out of memory.", .{});
-                    if (option.choices.?.len > 1 and i < option.choices.?.len - 1) choices_array.append(allocator, ", ") catch terminal.printError("Out of memory.", .{});
+                    choices_array.append(allocator, choice) catch terminal.printColored(&.{.red}, "Out of memory.", .{});
+                    if (option.choices.?.len > 1 and i < option.choices.?.len - 1) choices_array.append(allocator, ", ") catch terminal.printColored(&.{.red}, "Out of memory.", .{});
                 }
 
                 const choices = choices_array.toOwnedSlice(allocator) catch unreachable;
@@ -242,12 +242,12 @@ fn parseOptionValue(self: *Self, value: types.Value, option: CLIOption) types.Va
                 }
 
                 if (found == false) {
-                    terminal.printColored(.red, "Option '{s}' must be one of: [", .{option.name});
+                    terminal.printColored(&.{.red}, "Option '{s}' must be one of: [", .{option.name});
 
                     for (choices) |choice| {
-                        terminal.printColored(.red, "{s}", .{choice});
+                        terminal.printColored(&.{.red}, "{s}", .{choice});
                     }
-                    terminal.printColored(.red, "]", .{});
+                    terminal.printColored(&.{.red}, "]\n", .{});
                     std.process.exit(1);
                 }
             }
@@ -263,7 +263,7 @@ fn parseOptionValue(self: *Self, value: types.Value, option: CLIOption) types.Va
 //     for (args) |arg| {
 //         if (argIndex >= parsed.args.length) {
 //             if (arg.required) {
-//                 terminal.printColored(.red, "Missing required argument: {s}\n", .{arg.name});
+//                 terminal.printColored(&.{.red}, "Missing required argument: {s}\n", .{arg.name});
 //                 std.process.exit(1);
 //             }
 //             break;
@@ -291,7 +291,7 @@ pub fn validateOptions(self: *Self, parsed: ParsedArgs, options: []CLIOption) !P
 
     if (missing_options.items.len > 0) {
         const alias = if (missing_options.items[0].alias) |a| a else "";
-        terminal.printColored(.red, "Missing required option: --{s} (-{s})\n", .{ missing_options.items[0].name, alias });
+        terminal.printColored(&.{.red}, "Missing required option: --{s} (-{s})\n", .{ missing_options.items[0].name, alias });
         return error.MissingRequiredOption;
     }